+++ src/components/file_objects/character.rs
@@ -153,7 +153,7 @@ impl FileObject for Character {
 
     fn generate_outline(
         &self,
-        depth: u32,
+        depth: u64,
         export_string: &mut String,
         _objects: &super::FileObjectStore,
     ) {
